# api/openapi.yaml
# THIS FILE IS GENERATED FROM src/openapi/spec.ts. DO NOT EDIT DIRECTLY.
components:
  responses:
    Error:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
      description: エラーが発生した場合のレスポンス
  schemas:
    Bookmark:
      properties:
        createdAt: &a1
          description: ISO 8601 形式の日時
          format: date-time
          type: string
        id: &a2
          example: 1
          format: int64
          type: integer
        isRead: &a5
          type: boolean
        title: &a4
          nullable: true
          type: string
        updatedAt: *a1
        url: &a3
          format: uri
          type: string
      required:
        - id
        - url
        - isRead
        - createdAt
        - updatedAt
      type: object
    BookmarkListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - properties:
            bookmarks:
              items:
                $ref: "#/components/schemas/BookmarkWithFavorite"
              type: array
            todayReadCount:
              format: int64
              type: integer
            totalUnread:
              format: int64
              type: integer
          required:
            - bookmarks
          type: object
    BookmarkWithFavorite:
      properties:
        createdAt: *a1
        id: *a2
        isFavorite:
          type: boolean
        isRead: *a5
        title: *a4
        updatedAt: *a1
        url: *a3
      required:
        - id
        - url
        - isRead
        - createdAt
        - updatedAt
        - isFavorite
      type: object
    BulkBookmarksRequest:
      properties:
        bookmarks:
          items:
            properties:
              title:
                type: string
              url:
                format: uri
                type: string
            required:
              - url
              - title
            type: object
          minItems: 1
          type: array
      required:
        - bookmarks
      type: object
    ErrorResponse:
      properties:
        message:
          type: string
        success:
          enum:
            - false
          type: boolean
      required:
        - success
        - message
      type: object
    FavoriteBookmarksResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - properties:
            bookmarks:
              items:
                $ref: "#/components/schemas/BookmarkWithFavorite"
              type: array
            total:
              format: int64
              type: integer
          required:
            - bookmarks
          type: object
    MessageResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - properties:
            message:
              type: string
          required:
            - message
          type: object
    RecentBookmarksResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - properties:
            bookmarks:
              additionalProperties:
                items:
                  $ref: "#/components/schemas/BookmarkWithFavorite"
                type: array
              description: YYYY-MM-DD をキーにした既読ブックマークリスト
              type: object
          required:
            - bookmarks
          type: object
    SuccessResponse:
      properties:
        success:
          enum:
            - true
          type: boolean
      required:
        - success
      type: object
info:
  description: |-
    ブックマーク管理を行う Yomimono API の OpenAPI 定義。
    このファイルは src/openapi/spec.ts を唯一のソースとして生成されます。
  title: Yomimono API
  version: 0.1.0
openapi: 3.1.0
paths:
  /api/bookmarks:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkListResponse"
          description: 未読ブックマーク一覧
        "500":
          $ref: "#/components/responses/Error"
      summary: 未読ブックマーク一覧
      tags:
        - Bookmarks
  /api/bookmarks/bulk:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkBookmarksRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: 登録処理を実行
        "400":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
      summary: ブックマーク一括登録
      tags:
        - Bookmarks
  /api/bookmarks/favorites:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteBookmarksResponse"
          description: お気に入りのブックマーク一覧
        "500":
          $ref: "#/components/responses/Error"
      summary: お気に入り一覧
      tags:
        - Bookmarks
  /api/bookmarks/recent:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentBookmarksResponse"
          description: 日付ごとにグルーピングされた既読履歴
        "500":
          $ref: "#/components/responses/Error"
      summary: 最近既読にしたブックマーク
      tags:
        - Bookmarks
  /api/bookmarks/{id}/favorite:
    delete:
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
          description: お気に入りから削除
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
      summary: お気に入り削除
      tags:
        - Bookmarks
    post:
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
          description: お気に入り登録完了
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
      summary: お気に入り追加
      tags:
        - Bookmarks
  /api/bookmarks/{id}/read:
    patch:
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
          description: 既読化完了
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
      summary: ブックマークを既読にする
      tags:
        - Bookmarks
  /api/bookmarks/{id}/unread:
    patch:
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
          description: 未読へ戻した結果
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
      summary: ブックマークを未読に戻す
      tags:
        - Bookmarks
servers:
  - description: ローカル開発（wrangler dev）
    url: http://localhost:8787
  - description: 本番 API
    url: https://api-yomimono.workers.dev
tags:
  - description: ブックマーク関連の操作
    name: Bookmarks