# api/src/routes/bookmarks.ts
import { Hono } from "hono";
import {
	BadRequestError,
	ConflictError,
	createErrorResponse,
	createErrorResponseBody,
	InternalServerError,
	NotFoundError,
	toContentfulStatusCode,
} from "../exceptions";
import type { IBookmarkService } from "../interfaces/service/bookmark";
import { validateId } from "../utils/validation";

export const createBookmarksRouter = (bookmarkService: IBookmarkService) => {
	const app = new Hono();

	app.get("/", async (c) => {
		try {
			const [unreadBookmarks, totalUnread, todayReadCount] = await Promise.all([
				bookmarkService.getUnreadBookmarks(),
				bookmarkService.getUnreadBookmarksCount(),
				bookmarkService.getTodayReadCount(),
			]);
			return c.json({
				success: true,
				bookmarks: unreadBookmarks,
				totalUnread,
				todayReadCount,
			});
		} catch (error) {
			console.error("Failed to fetch bookmarks:", error);
			const errorResponse = createErrorResponse(error);
			return c.json(
				createErrorResponseBody(error),
				toContentfulStatusCode(errorResponse.statusCode),
			);
		}
	});

	app.post("/bulk", async (c) => {
		try {
			const { bookmarks } = await c.req.json<{
				bookmarks: Array<{ url: string; title: string }>;
			}>();

			if (!Array.isArray(bookmarks)) {
				throw new BadRequestError("bookmarks must be an array");
			}

			if (bookmarks.length === 0) {
				throw new BadRequestError("bookmarks array cannot be empty");
			}

			// URLの形式チェック
			const isValidUrl = (url: string) => {
				try {
					new URL(url);
					return true;
				} catch {
					return false;
				}
			};

			if (!bookmarks.every((b) => isValidUrl(b.url))) {
				throw new BadRequestError("invalid URL format");
			}

			const totalCount = bookmarks.length;
			await bookmarkService.createBookmarksFromData(bookmarks);
			return c.json({
				success: true,
				message: `Processed ${totalCount} bookmarks (duplicates skipped if unread)`,
			});
		} catch (error) {
			console.error("Failed to create bookmarks:", error);
			const errorResponse = createErrorResponse(error);
			return c.json(
				createErrorResponseBody(error),
				toContentfulStatusCode(errorResponse.statusCode),
			);
		}
	});

	// お気に入り機能のルーティングを追加
	app.post("/:id/favorite", async (c) => {
		try {
			const id = validateId(c.req.param("id"), "bookmark ID");

			await bookmarkService.addToFavorites(id);
			return c.json({ success: true });
		} catch (error) {
			if (error instanceof NotFoundError) {
				return c.json(createErrorResponseBody(error), 404);
			}
			if (error instanceof ConflictError) {
				return c.json(createErrorResponseBody(error), 409);
			}
			if (error instanceof BadRequestError) {
				return c.json(createErrorResponseBody(error), 400);
			}
			console.error("Failed to add to favorites:", error);
			const internalError = new InternalServerError(
				"Failed to add to favorites",
			);
			return c.json(createErrorResponseBody(internalError), 500);
		}
	});

	app.delete("/:id/favorite", async (c) => {
		try {
			const id = validateId(c.req.param("id"), "bookmark ID");

			await bookmarkService.removeFromFavorites(id);
			return c.json({ success: true });
		} catch (error) {
			if (error instanceof NotFoundError) {
				return c.json(createErrorResponseBody(error), 404);
			}
			if (error instanceof BadRequestError) {
				return c.json(createErrorResponseBody(error), 400);
			}
			console.error("Failed to remove from favorites:", error);
			const internalError = new InternalServerError(
				"Failed to remove from favorites",
			);
			return c.json(createErrorResponseBody(internalError), 500);
		}
	});

	app.get("/favorites", async (c) => {
		try {
			const result = await bookmarkService.getFavoriteBookmarks();
			return c.json({ success: true, ...result });
		} catch (error) {
			console.error("Failed to fetch favorites:", error);
			const errorResponse = createErrorResponse(error);
			return c.json(
				createErrorResponseBody(error),
				toContentfulStatusCode(errorResponse.statusCode),
			);
		}
	});

	// 既読機能
	app.patch("/:id/read", async (c) => {
		try {
			const id = validateId(c.req.param("id"), "bookmark ID");

			await bookmarkService.markBookmarkAsRead(id);
			return c.json({ success: true });
		} catch (error) {
			if (error instanceof NotFoundError) {
				return c.json(createErrorResponseBody(error), 404);
			}
			if (error instanceof BadRequestError) {
				return c.json(createErrorResponseBody(error), 400);
			}
			console.error("Failed to mark bookmark as read:", error);
			const internalError = new InternalServerError(
				"Failed to mark bookmark as read",
			);
			return c.json(createErrorResponseBody(internalError), 500);
		}
	});

	// 未読に戻す機能
	app.patch("/:id/unread", async (c) => {
		try {
			const id = validateId(c.req.param("id"), "bookmark ID");

			await bookmarkService.markBookmarkAsUnread(id);
			return c.json({ success: true });
		} catch (error) {
			if (error instanceof NotFoundError) {
				return c.json(createErrorResponseBody(error), 404);
			}
			if (error instanceof BadRequestError) {
				return c.json(createErrorResponseBody(error), 400);
			}
			console.error("Failed to mark bookmark as unread:", error);
			const internalError = new InternalServerError(
				"Failed to mark bookmark as unread",
			);
			return c.json(createErrorResponseBody(internalError), 500);
		}
	});

	app.get("/recent", async (c) => {
		try {
			const recentlyReadBookmarks =
				await bookmarkService.getRecentlyReadBookmarks();
			return c.json({ success: true, bookmarks: recentlyReadBookmarks });
		} catch (error) {
			console.error("Failed to fetch recently read bookmarks:", error);
			const errorResponse = createErrorResponse(error);
			return c.json(
				createErrorResponseBody(error),
				toContentfulStatusCode(errorResponse.statusCode),
			);
		}
	});

	return app;
};